requirements.txt
fastapi==0.115.6
numpy==2.2.1
pyaudio==0.2.14
python-dotenv==1.0.1
scipy==1.15.0
uvicorn==0.34.0
websockets==14.1
aiohttp==3.11.11
